+++ yogurt_editor/lib/src/views/cell_view.dart
@@ -51,25 +51,23 @@ class _CellViewState extends State<CellView> {
     _subscription = null;
   }
 
-  Widget _build(BuildContext context) {
-    Widget content = _state.model.build(context, _state);
-    for (var plugin in widget.controller.plugins.reversed) {
-      content = plugin.build(context, _controller, content);
-    }
-
-    return content;
-  }
-
   @override
   Widget build(BuildContext context) {
-    return Stack(
+    Widget content = Stack(
+      fit: StackFit.expand,
       children: [
-        _build(context),
+        _state.model.build(context, _state),
         for (final child in _controller.children.values)
           CellView(
             controller: child,
           ),
       ],
     );
+
+    for (var plugin in widget.controller.plugins.reversed) {
+      content = plugin.build(context, _controller, content);
+    }
+
+    return content;
   }
 }