+++ yogurt_editor/CHANGELOG.md
@@ -1,11 +1,11 @@
 ## 0.0.3
 
-* Refactor.
+- fix: child cell should be effected by parent cell plugin
 
 ## 0.0.2
 
-* Refactor.
+- Refactor.
 
 ## 0.0.1
 
-* First version.
+- First version.